Extract child-lookup for consistent reuse
