+++ app/models/group/regionalverband.rb
@@ -54,16 +54,7 @@ class Group::Regionalverband < ::Group
            Group::RegionalverbandKontakte,
            Group::Verein
 
-
-  def children
-    primary = super
-    secondary = ::Group::Verein.where(secondary_parent_id: id)
-    tertiary  = ::Group::Verein.where(tertiary_parent_id: id)
-
-    ids = (primary | secondary | tertiary).map(&:id)
-
-    Group.where(id: ids)
-  end
+  include SecondaryChildren
 
   ### ROLES
